requirements.txt
flask>=3.0
openai>=1.0
boto3>=1.35
paramiko>=3.0
runpod>=1.7
verda>=1.20
python-dotenv>=1.0
